requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
